Adding restrictions, company dashboard plus created jobs

+++ controllers/GigsController.js
@@ -2,8 +2,9 @@ const Gig = require('../models/Gig');
 const Sequilize = require('sequelize');
 const User = require('../models/User');
 const UserGigs = require('../models/UserGigs');
-
+const Roles = require('../utils/Roles');
 const Op = Sequilize.Op;
+const Company = require('../models/Company');
 
 const log = console.log;
 
@@ -14,12 +15,18 @@ exports.getAllGigs = async(req,res,next)=>{
     try{
         const gigs = await Gig.findAll( {order: [ [ 'createdAt', 'DESC' ]]});
         let msg = req.flash('message').slice();
-
+        let user,company;
+        
+        if(req.user){
+            user = req.user.role == Roles.USER ? req.user: null;
+            company =   req.user.role == Roles.COMPANY ? req.user : null ;
+        }
         res.render('gigs',{
             status:'success',
             message: msg,
             gigs:gigs,
-            user:req.user
+            user:user,
+            company:company
         })
     }catch(err){
         console.log(err)
@@ -30,7 +37,12 @@ exports.getAllGigs = async(req,res,next)=>{
 // @route  GET /gigs/add 
 // @acces   Company (in the future)
 exports.createNewGig = async (req,res,next)=>{
-    res.render('add',{user:req.user});
+    let user,company;
+    if(req.user){
+        user = req.user.role == Roles.USER ? req.user: null;
+        company =   req.user.role == Roles.COMPANY ? req.user : null ;
+    }
+    res.render('add',{user:user,company:company});
 }
 
 
@@ -38,7 +50,12 @@ exports.createNewGig = async (req,res,next)=>{
 // @route  POST /gigs/add 
 // @acces   Company (in the future)
 exports.storeNewGig = async (req,res,next)=>{
-    
+    let user,company;
+    if(req.user){
+        user = req.user.role == Roles.USER ? req.user: null;
+        company =   req.user.role == Roles.COMPANY ? req.user : null ;
+    }
+
     let {title,technologies, budget,description,contact_email} = req.body;
     let errors = [];
 
@@ -65,7 +82,9 @@ exports.storeNewGig = async (req,res,next)=>{
              budget,
              description,
              contact_email,
-             user:req.user,
+             user:user,
+             company:company
+      
         })
     }else{
 
@@ -89,7 +108,8 @@ exports.storeNewGig = async (req,res,next)=>{
             technologies,
             budget,
             description,
-            contact_email
+            contact_email,
+            companyId:req.user.id
         })
         .then(gig=>{
             req.flash('message','You have created a gig successfully!')
@@ -107,9 +127,13 @@ exports.storeNewGig = async (req,res,next)=>{
 exports.searchForGigs = async (req,res,next)=>{
     const term = req.query.term;
     let lowerCaseterm = term.toLowerCase();
-
+    let user,company;
+    if(req.user){
+        user = req.user.role == Roles.USER ? req.user: null;
+        company =   req.user.role == Roles.COMPANY ? req.user : null ;
+    }
     Gig.findAll({where:{technologies:{[Op.like]: '%'+ lowerCaseterm +'%'}}})
-    .then(gigs=>res.render('gigs',{gigs:gigs,user:req.user}))
+    .then(gigs=>res.render('gigs',{gigs:gigs,user:user,company:company}))
     .catch(err => console.log(err));
 
 }
@@ -132,30 +156,38 @@ exports.gigDetails = async (req,res,next)=>{
 // @route   Post /gigs/delete/:id 
 // @acces   Company
 exports.destroyGig = async (req,res,next)=>{
-
+    let user,company;
+    if(req.user){
+        user = req.user.role == Roles.USER ? req.user: null;
+        company =   req.user.role == Roles.COMPANY ? req.user : null ;
+    }
     const {id} = req.params;
-    UserGigs.findAll({
-        where:{
-            gigId:id
+    if(company){
+        const owned = await gigOwnedByCompany(company,id);
+        if(owned){
+            UserGigs.findAll({
+                where:{
+                    gigId:id
+                }
+            }).then(results=>{
+               results.forEach(result=>{
+                   result.destroy();
+               })
+            })
+            .then(()=>{
+                Gig.destroy({
+                    where:{id:id}
+                })
+                .then(gig=>{
+                    req.flash('message','Gig deleted successfully!')
+                    res.redirect('/gigs');
+                })
+                .catch(err => console.log(err));
+            })
+            .catch(err=>console.log(err));
         }
-    }).then(results=>{
-       results.forEach(result=>{
-           result.destroy();
-       })
-    })
-    .then(()=>{
-        Gig.destroy({
-            where:{id:id}
-        })
-        .then(gig=>{
-           
-            const message = `${gig.title} has been destroyed`;
-            req.flash('message','Gig deleted successfully!')
-            res.redirect('/gigs');
-        })
-        .catch(err => console.log(err));
-    })
-    .catch(err=>console.log(err));
+    }
+    
 
 }
 
@@ -180,4 +212,23 @@ function alreadyApplied(userId,gigId){
   return  UserGigs.count({where:{userId:userId,gigId:gigId}})
     .then(count => count !=0 ?  true :  false )
     .catch(err=>console.log(err));
+}
+
+// Check if the Gig is owned by the company
+async function gigOwnedByCompany(company,gigId){
+    let foundJobs= [];
+       try{
+            const comp = await Company.findAll({
+                where:{id:company.id},
+                include:[Gig],
+                });
+            comp[0].gigs.forEach(gig=>{
+                foundJobs.push(gig.get());
+            })
+         let hasTheJob = foundJobs.filter(x=>x.id==gigId);
+         return hasTheJob.length > 0;
+       }
+       catch(err){
+           log(err);
+       }
 }
